+++ code/src/components/AddToDoTest/AddToDo.js
@@ -11,7 +11,7 @@ import {
   AddContainer,
   Select,
   Option
-} from 'components/AddToDo/AddToDo.style'
+} from 'components/AddToDoTest/AddToDo.style'
 
 export const AddToDo = () => {
   const [inputValue, setInputValue] = useState('');
